Separate bson codecs from BsonObject
