--- mewbase-core/src/main/java/io/mewbase/bson/JsonValueToBsonValue.java
@@ -1,4 +1,0 @@
-package io.mewbase.bson;
-
-public class JsonValueToBsonValue {
-}
