+++ mewbase-core/src/test/java/io/mewbase/bson/BsonObjectTest.java
@@ -21,11 +21,13 @@ package io.mewbase.bson;
 import io.mewbase.TestUtils;
 
 import io.vertx.core.buffer.Buffer;
+import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonObject;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.io.StringReader;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
@@ -53,13 +55,12 @@ public class BsonObjectTest {
         final String key = "event_type";
         final String value = "manually_typed_address_merged";
         final String jsonStr = "{\""+key+"\":\""+value+"\"}";
+        final StringReader stringReader = new StringReader(jsonStr);
         final JsonObject jsonObject = new JsonObject(jsonStr);
         final BsonObject bsonObject = new BsonObject(jsonObject);
         assertNotNull(bsonObject);
         assertEquals(bsonObject.getString(key),value);
     }
-
-
 
     @Test
     public void testGetInteger() {
